pyproject.toml
[project]
name = "catch-knowledge"
version = "0.1.0"
description = "Personal interview experience knowledge base pipeline."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apscheduler>=3.10,<4.0",
    "beautifulsoup4>=4.12,<5.0",
    "httpx>=0.27,<0.28",
    "openai>=1.30,<2.0",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "python-dateutil>=2.9,<3.0",
    "sqlalchemy>=2.0,<2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2,<9.0",
    "ruff>=0.4,<0.5",
]
playwright = [
    "playwright>=1.44,<2.0",
]
postgres = [
    "psycopg[binary]>=3.1,<3.3",
]

[project.scripts]
catch-knowledge = "catch_knowledge.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
